general revision of the code

+++ exercice_2.15.5.py
@@ -1,3 +1,3 @@
-temperature = int(input('Entrez la temperature en celsius: '))
+temperature = float(input('Entrez la temperature en celsius: '))
 farh = (temperature*9/5)+32
 print(farh)
